Refactored everything from meta to App\Domain\Meta namespace
