+++ composer.json
@@ -78,6 +78,7 @@
             "Console\\": "src/Console",
             "ForkCMS\\": "src/ForkCMS",
             "ForkCMS\\App\\": "app",
+            "App\\": "src",
             "CKSource\\CKFinder\\": "src/Backend/Core/Js/ckfinder/core/connector/php/CKSource/CKFinder"
         },
         "exclude-from-classmap": [
