+++ src/Backend/Modules/Faq/Domain/Question/Question.php
@@ -3,7 +3,7 @@
 namespace Backend\Modules\Faq\Domain\Question;
 
 use Backend\Modules\Faq\Domain\Category\Category;
-use Common\Doctrine\Entity\Meta;
+use App\Domain\Meta\Meta;
 use Common\Locale;
 use DateTime;
 use Doctrine\Common\Collections\Collection;
@@ -35,7 +35,7 @@ class Question
     /**
      * @var Meta
      *
-     * @ORM\OneToOne(targetEntity="Common\Doctrine\Entity\Meta", cascade={"persist","remove"}, orphanRemoval=true)
+     * @ORM\OneToOne(targetEntity="App\Domain\Meta\Meta", cascade={"persist","remove"}, orphanRemoval=true)
      */
     private $meta;
