+++ src/Frontend/Core/Engine/Base/Block.php
@@ -3,7 +3,7 @@
 namespace Frontend\Core\Engine\Base;
 
 use Common\Core\Header\Priority;
-use Common\Doctrine\Entity\Meta;
+use App\Domain\Meta\Meta;
 use Common\Exception\RedirectException;
 use ForkCMS\App\KernelLoader;
 use Frontend\Core\Engine\Breadcrumb;
